requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
docutils==0.20.1
glcontext==2.5.0
idna==3.4
jax==0.4.20
jaxlib==0.4.20
joblib==1.3.2
Kivy==2.2.1
Kivy-examples==2.2.1
Kivy-Garden==0.1.5
ml-dtypes==0.3.1
moderngl==5.9.0
numpy==1.26.2
opt-einsum==3.3.0
Pillow==10.1.0
pyassimp==5.2.5
pygame==2.5.0
Pygments==2.17.2
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
pyopengltk==0.0.4
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
threadpoolctl==3.2.0
urllib3==2.1.0
